% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mndot}
\alias{mndot}
\title{Minidot import}
\usage{
mndot(file_path)
}
\arguments{
\item{file_path}{Characther path to the file}
}
\description{
Import directly minidot files concatenated with all its columns ready to use for any purpose
The package includes minidots files for trial
}
