% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_data.R
\name{use_data}
\alias{use_data}
\title{usedata::use_data calls usethis::use_data and usethis::use_r}
\usage{
use_data(
  ...,
  title = "dataset",
  description = "dataset of observations for something",
  source = "data source comes from",
  internal = FALSE,
  overwrite = TRUE,
  compress = "bzip2",
  version = 3,
  ascii = FALSE
)
}
\arguments{
\item{...}{a single dataset you want to document}

\item{title}{a title to give for the doc script}

\item{description}{a description of your data for the doc script}

\item{source}{the source of the dataset}

\item{internal}{true or false passes to usethis::use_data}

\item{overwrite}{true or false passes to usethis::use_data}

\item{compress}{compression for usethis::use_data}

\item{version}{version defaults to 3 for usethis::use_data}

\item{ascii}{ascii true or false for usethis::use_data}
}
\value{
saves off data in the pkg then creates and fills out a R script doc
}
\description{
usedata::use_data calls usethis::use_data and usethis::use_r
}
