% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usedata-package.R
\docType{package}
\name{usedata-package}
\alias{usedata}
\alias{usedata-package}
\title{usedata: built with "usethis" to help document package data}
\description{
usedata has a single goal, to help R package builders add and document data efficiently. usedata is based on R Pkg 2e Chapter 7. automating small manual steps for developers that register lots of example datasets.
}
\author{
\strong{Maintainer}: frank hull \email{frankiethull@gmail.com}

}
\keyword{internal}
