% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mndot}
\alias{mndot}
\title{Minidot function}
\usage{
mndot(file_choose = F, path)
}
\arguments{
\item{file_chose}{Interactively chose the file that you need to import, if FALSE write the path}
}
\description{
Import directly minidot files concatenated with all its columns ready to use for any purpose
}
